Allow for library suffixes
Linux installations like Slackware64 use the lib64 prefix to
distinguish 64-bit libraries from 32-bit installations. We
should enable support for this in the Makefile through config.Mk.

Unfortunately, this does not make things perfect, since there are
some scripts that assume we use lib instead of lib${LIBSUFFIX} and
we need to have some way of handling that.