Add rudementary word wrap support in place of wrap to fit
In lieu of having an authentic wrap to fit feature in the system,
I've disabled the modify callback that attempted to do that
previously in favor of just using the built in Motif Text wordWrap
resource.

This gives us a well tested word wrap feature, but it doesn't quite
match the traditional approach. I think that this will work, though
it is a bit annoying.